libdnf/logger: Add a StringLogger base class
A base class for loggers logging as lines of text (non-structured
loggers). This class allows to share the log line formatting code.
